requirements.txt
minigrid>=2.3.1
gymnasium>=0.29.0
numpy>=1.24.0
tensorboard>=2.15.0
pyyaml>=6.0
matplotlib>=3.7.0
